requirements.txt
asgiref==3.2.7
Django==3.0.5
django-environ==0.4.5
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.1
mysqlclient==1.4.6
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pytz==2019.3
sqlparse==0.3.1
